requirements.txt
#json
numpy
pillow
tensorflow
tqdm